Authorization-part-one

+++ src/main/java/dev/tolana/projectcalculationtool/dto/UserEntityRoleDto.java
@@ -1,12 +1,12 @@
 package dev.tolana.projectcalculationtool.dto;
 
 public record UserEntityRoleDto(String username,
-                                int roleId,
-                                int taskId,
-                                int projectId,
-                                int teamId,
-                                int departmentId,
-                                int organizationId) {
+                                long roleId,
+                                long taskId,
+                                long projectId,
+                                long teamId,
+                                long departmentId,
+                                long organizationId) {
 }
